Fixed A Diagonal + Progress + Why?
So, I've had these changes just sitting on my computer for months (since
the previous commit - I believe I was pretty active before dropping the
project), and I finally decided to do something about that.

Most of this code was already done months ago, but lines 739-771 were
added by me today. Also, all the comments on the *new* code (not
modified lines, but the NEW lines) didn't exist until today... Yes, I'm
serious.

Anyway, while messing around I also found a bug in the game logic that I
had never noticed before. If you try to win with a top-right to
bottom-left diagonal pattern, and the final piece to place to finish
this line, is the one in the center of the board, and you have a piece
in the top left (but not the bottom right!), then you wouldn't win.

You'll notice in the testWin procedure, I commented out one line. That's
the issue :) Basically the check goes: there is a piece in the top-left,
okay, now check the bottom-right. We don't have a piece there? Okay,
then they didn't win.

It works fine if you don't have a piece in the top left, since I made
sure to then jump down to the part of the code that was ALREADY WRITTEN
to check for a top-right to bottom-left diagonal win, but yeah, the
problem is that I incorrectly jumped to the end of the routine without
checking that in certain situations.

This work on the help screen isn't finished, for the record. But seems
to "work" despite a gnawing feeling that the code I wrote is very
terrible and slow and bad and I hate it.